pyproject.toml
[project]
name = "write_a_book_with_flows"
version = "0.1.0"
description = "write_a_book_with_flows using crewAI"
authors = [
    { name = "Your Name", email = "you@example.com" },
]
requires-python = ">=3.10,<=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "crewai[tools]==0.85.0",
    "google-generativeai==0.8.4",
    "streamlit==1.43.1",
    "python-docx==1.1.2",

]

[project.scripts]
kickoff = "write_a_book_with_flows.web.app:kickoff"
plot = "write_a_book_with_flows.main:plot"

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"
